LAB_1_PROCUREMENT_AGENT/backup/BE_google_sheet/requirements.txt
requests
fastapi
uvicorn[standard]
python-dotenv
pydantic>=2.0
gspread
google-auth
